Fix UI styling content view
