+++ frontend/src/app/styles/content/viewContent.scss
@@ -7,11 +7,35 @@
   -webkit-backdrop-filter: blur(10px);
 }
 
+.content-title-bar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 20px;
+}
+
+.content-interactions {
+  display: flex;
+  gap: 5px;
+}
+
+.content-interactions .icon {
+  width: 25px;
+  height: 25px;
+}
+
 .content-header {
   display: flex;
   align-items: center;
   gap: 15px;
   margin-bottom: 40px;
+}
+
+.profile-image-creator {
+  border-radius: 50%;
+  object-fit: cover;
+
+  cursor: pointer;
 }
 
 .row {
@@ -33,36 +57,48 @@
   width: 70%;
 }
 
+/* Icon styles */
+.icon {
+  width: 16px;
+  height: 16px;
+  color: var(--text-color);
+  transition: all ease-out 0.3s;
+}
+
+.icon.cancel {
+  width: 25px;
+  height: 25px;
+}
+
+.icon:hover {
+  color: var(--icon-hover-color);
+}
+
+.icon.delete:hover {
+  color: var(--delete-hover-color);
+}
+
+.icon.cancel:hover {
+  color: var(--cancel-hover-color);
+}
+
+.icon.counter {
+  font-size: 12px;
+  margin-top: 0;
+  width: fit-content;
+  display: none;
+}
+
+.icon.counter.visible {
+  display: inline;
+}
+
 .thumbnail {
   width: 100%;
   height: 300px;
   border-radius: 10px;
   object-fit: cover;
   background-color: var(--tile-color);
-}
-
-.username-follow {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-}
-
-.icon-button.follow {
-  padding: 0;
-  border: none;
-  background: none;
-  display: inline-flex;
-  align-items: center;
-}
-
-.icon.follow {
-  width: 16px;
-  height: 16px;
-  color: var(--icon-color);
-}
-
-.icon.follow.following {
-  color: black;
 }
 
 .icon-container {
@@ -79,10 +115,36 @@
   align-items: center;
   position: relative;
   gap: 1px;
+  cursor: pointer;
 }
 
+.creator-follow-section {
+  position: relative;
+}
+
+.follow-button {
+  position: absolute;
+  bottom: -5px;
+  left: 40px;
+
+  padding: 5px 10px;
+  border-radius: 10px;
+
+  background-color: var(--tile-color);
+  color: var(--text-color);
+  border: 1px solid var(--input-border-color);
+
+  cursor: pointer;
+}
+
+.follow-button:hover {
+  background-color: var(--input-border-color);
+  color: var(--background-color);
+}
+
+// COMMENT SECTION
 .add-button {
-  color: var(--icon-color);
+  color: var(--text-color);
   transition: all ease-out 0.3s;
   width: 30px;
   position: absolute;
@@ -107,50 +169,6 @@
   flex-direction: column-reverse;
   text-align: right;
   gap: 0.3rem;
-}
-
-/* Icon styles */
-.icon {
-  width: 16px;
-  height: 16px;
-  color: var(--icon-color);
-  transition: all ease-out 0.3s;
-}
-
-.icon.cancel {
-  width: 25px;
-  height: 25px;
-}
-
-.icon:hover {
-  color: var(--icon-hover-color);
-}
-
-.icon.liked {
-  color: black;
-}
-
-.icon.bookmarked {
-  color: black;
-}
-
-.icon.delete:hover {
-  color: var(--delete-hover-color);
-}
-
-.icon.cancel:hover {
-  color: var(--cancel-hover-color);
-}
-
-.icon.counter {
-  font-size: 12px;
-  margin-top: 0;
-  width: fit-content;
-  display: none;
-}
-
-.icon.counter.visible {
-  display: inline;
 }
 
 .comment-textarea {
